extract StreamToLoggerAdaptor
