+++ core/amber/src/main/python/core/util/print_writer/stream_to_logger_adaptor.py
@@ -1,0 +1,17 @@
+from loguru import logger
+
+
+class StreamToLoggerAdaptor:
+    """
+    This class is used to redirect `print` to loguru's logger, instead of stdout.
+    """
+
+    def __init__(self, level=logger.level("PRINT", no=38)):
+        self._level = level
+
+    def write(self, buffer):
+        for line in buffer.rstrip().splitlines():
+            logger.opt(depth=1).log("PRINT", line.rstrip())
+
+    def flush(self):
+        pass
