+++ core/amber/src/main/python/texera_run_python_worker.py
@@ -1,28 +1,10 @@
 import contextlib
 import sys
 
-from loguru import logger
-
 from core.python_worker import PythonWorker
-
-
-class StreamToLogger:
-    """
-    This class is used to redirect `print` to loguru's logger, instead of stdout.
-    """
-
-    def __init__(self, level=logger.level("PRINT", no=38)):
-        self._level = level
-
-    def write(self, buffer):
-        for line in buffer.rstrip().splitlines():
-            logger.opt(depth=1).log("PRINT", line.rstrip())
-
-    def flush(self):
-        pass
-
+from core.util.print_writer.stream_to_logger_adaptor import StreamToLoggerAdaptor
 
 if __name__ == '__main__':
     # redirect user's print into logger
-    with contextlib.redirect_stdout(StreamToLogger()):
+    with contextlib.redirect_stdout(StreamToLoggerAdaptor()):
         PythonWorker(host="localhost", input_port=int(sys.argv[1]), output_port=int(sys.argv[2])).run()
